Extract shared MemZip in-memory backend, migrate PPTExporter, and exclude it from the WeChat build.
